Use `ag` only via Ack.vim to avoid slowdown
Something about the way we were running The Silver Searcher as a direct
vim command, i.e., `:Ag blah` was causing serious performance issues if
that search happened to find a match inside a huge file (like a JSON
dump). This problem persists even after the quick results window is
closed; restarting Vim appears to be the only remedy.

Fortunately, this is not a problem when the search is conducted via the
Ack.vim plugin (even when `ag` is the used as the backend instead of
`ack` proper.