Let CtrlP use default search method
Also, add some ignores. It's better to use the default than The Silver
Searcher; the latter tends to bog down when you are in a large directory
(like ~). Caching isn't so bad since it only takes a couple of seconds
even on a large dir & subsequent accesses to C-p are instantaneous.